# docs/mall_admin/mall_admin_swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "mall-go API",
        "title": "mall-go API",
        "contact": {},
        "version": "v1"
    },
    "basePath": "/",
    "paths": {
...
        "/resource/create": {
            "post": {
                "security": [
                    {
                        "GinJWTMiddleware": []
                    }
                ],
                "description": "添加后台资源",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台资源管理"
                ],
                "summary": "添加后台资源",
                "parameters": [
                    {
                        "description": "添加后台资源",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ums_admin.UmsResourceCreateDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin_common.GinCommonResponse"
                        }
                    }
                }
            }
        },
        "/resource/delete/{id}": {
            "post": {
                "security": [
                    {
                        "GinJWTMiddleware": []
                    }
                ],
                "description": "根据ID删除后台资源",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台资源管理"
                ],
                "summary": "根据ID删除后台资源",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "资源id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin_common.GinCommonResponse"
                        }
                    }
                }
            }
        },
        "/resource/list/resource": {
            "get": {
                "security": [
                    {
                        "GinJWTMiddleware": []
                    }
                ],
                "description": "分页模糊查询后台资源",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台资源管理"
                ],
                "summary": "分页模糊查询后台资源",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "资源分类ID",
                        "name": "categoryId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "资源名称",
                        "name": "nameKeyword",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "资源URL",
                        "name": "urlKeyword",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "pageNum",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin_common.GinCommonResponse"
                        }
                    }
                }
            }
        },
        "/resource/listAll": {
            "get": {
                "security": [
                    {
                        "GinJWTMiddleware": []
                    }
                ],
                "description": "查询所有后台资源",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台资源管理"
                ],
                "summary": "查询所有后台资源",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin_common.GinCommonResponse"
                        }
                    }
                }
            }
        },
        "/resource/update/{id}": {
            "post": {
                "security": [
                    {
                        "GinJWTMiddleware": []
                    }
                ],
                "description": "修改后台资源",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台资源管理"
                ],
                "summary": "修改后台资源",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "资源id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "修改后台资源",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ums_admin.UmsResourceCreateDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin_common.GinCommonResponse"
                        }
                    }
                }
            }
        },
        "/resource/{id}": {
            "get": {
                "security": [
                    {
                        "GinJWTMiddleware": []
                    }
                ],
                "description": "根据ID获取资源详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台资源管理"
                ],
                "summary": "根据ID获取资源详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "资源id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin_common.GinCommonResponse"
                        }
                    }
                }
            }
        },
        "/resourceCategory/create": {
            "post": {
                "security": [
                    {
                        "GinJWTMiddleware": []
                    }
                ],
                "description": "添加后台资源分类",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台资源分类管理"
                ],
                "summary": "添加后台资源分类",
                "parameters": [
                    {
                        "description": "添加后台资源分类",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ums_admin.UmsResourceCategoryCreateDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin_common.GinCommonResponse"
                        }
                    }
                }
            }
        },
        "/resourceCategory/delete/{id}": {
            "post": {
                "security": [
                    {
                        "GinJWTMiddleware": []
                    }
                ],
                "description": "删除后台资源分类",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台资源分类管理"
                ],
                "summary": "删除后台资源分类",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin_common.GinCommonResponse"
                        }
                    }
                }
            }
        },
        "/resourceCategory/listAll": {
            "get": {
                "security": [
                    {
                        "GinJWTMiddleware": []
                    }
                ],
                "description": "查询所有后台资源分类",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台资源分类管理"
                ],
                "summary": "查询所有后台资源分类",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin_common.GinCommonResponse"
                        }
                    }
                }
            }
        },
        "/resourceCategory/update/{id}": {
            "post": {
                "security": [
                    {
                        "GinJWTMiddleware": []
                    }
                ],
                "description": "修改后台资源分类",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台资源分类管理"
                ],
                "summary": "修改后台资源分类",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "修改后台资源分类",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ums_admin.UmsResourceCategoryCreateDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin_common.GinCommonResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "gin_common.GinCommonResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "http 状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "返回的数据是任意类型\t如果有错误，则把错误信息也封装在此\n\n\t\t{\n\t\t\t\"err_code\": 300000,\n\t\t\t\"err_msg\": \"用户名已存在\"\n\t\t}"
                },
                "message": {
                    "description": "返回的信息 例如: 操作成功或者操作失败",
                    "type": "string"
                },
                "status": {
                    "description": "success or fail",
                    "type": "string"
                }
            }
        },
...
        "ums_admin.UmsResourceCategoryCreateDTO": {
            "type": "object",
            "properties": {
                "name": {
                    "description": "资源分类名称",
                    "type": "string"
                },
                "sort": {
                    "description": "排序",
                    "type": "integer"
                }
            }
        },
        "ums_admin.UmsResourceCreateDTO": {
            "type": "object",
            "properties": {
                "categoryId": {
                    "description": "资源分类ID",
                    "type": "integer"
                },
                "description": {
                    "description": "描述",
                    "type": "string"
                },
                "name": {
                    "description": "资源名称",
                    "type": "string"
                },
                "url": {
                    "description": "资源URL",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "GinJWTMiddleware": {
            "description": "Description for what is this security definition being used",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}